refactor: expose collector as a Node function too
