+++ lib/setup-beacon-recording.js
@@ -11,7 +11,7 @@ const { safeJSONParse } = require('./tools');
 const fs = require('fs');
 const url = require('url');
 const path = require('path');
-const logger = require('./logger');
+const logger = require('./logger')();
 
 const { PuppeteerBlocker, fromPuppeteerDetails } = require('@cliqz/adblocker-puppeteer');
