+++ lib/setup-cookie-recording.js
@@ -10,7 +10,7 @@
 const { safeJSONParse } = require('./tools');
 const cookieParser = require('tough-cookie').Cookie;
 const fs = require('fs');
-const logger = require('./logger');
+const logger = require('./logger')();
 const url = require('url');
 const groupBy = require('lodash/groupBy');
