+++ lib/setup-websocket-recording.js
@@ -8,7 +8,7 @@
 // jshint esversion: 8
 
 const { safeJSONParse } = require('./tools');
-const logger = require('./logger');
+const logger = require('./logger')();
 
 module.exports.setup_websocket_recording = function(page) {
   // recording websockets
