+++ lib/tools.js
@@ -9,7 +9,7 @@
 
 const url = require('url');
 const path = require ('path');
-const logger = require('./logger');
+const logger = require('./logger')();
 
 safeJSONParse = function (obj) {
   try {
